# [win32] ImageDataProvider getBounds performance