@HeikoKlare 2025-09-30
For the presented scenario the change definitely improves performance. And it of course makes sense, because when only the bounds for a zoom are of interest, there is no need to create the _right_ image data (via handle creation) that are properly _scaled_. It is sufficient to just find any representative image data from which we can derive the correct bounds. In my understanding, the proposed change is correct w.r.t. that regard.

I only have one concern: I would expect that when retreiving the bounds for a specific zoom, the consumer will likely also request image data or even a handle for that zoom later on. And in that scenario, the change leads to a performance drawback. Take the following snippet (derived from the on in the PR description):
```java
	public static void main(String[] args) {
		Display display = new Display();
		ImageDataProvider idp = new ImageDataProvider() {
			@Override
			public ImageData getImageData(int zoom) {
				ImageData data = new ImageData(2000, 2000, 32, new PaletteData(0, 0, 0));
				for (int x = 0; x < 2000; x++) {
					for (int y = 0; y < 2000; y++) {
						data.setPixel(x, y, 1);
					}
				}
				return data;
			}
		};
		long start = System.currentTimeMillis();
		for (int i = 0; i < 100; i++) {
			Image image = new Image(display, idp);
			image.getBounds();
			image.getImageData();
		}
		System.out.println("Duration ImageDataProvider " +  (System.currentTimeMillis() - start) + " ms size: " + new Image(display, idp).getBounds());
	}
```

On average, I get the following results:\
Without PR: Duration ImageDataProvider 6350 ms size: Rectangle {0, 0, 2000, 2000} \
With PR: Duration ImageDataProvider 7900 ms size: Rectangle {0, 0, 2000, 2000}